requirements.txt
Django>=6.0,<7.0
google-generativeai
pillow
python-dotenv
